Export levenshtein suggestion utility package
