+++ pkg/edition/java/proxy/builtin_cmd_glist.go
@@ -2,13 +2,14 @@ package proxy
 
 import (
 	"fmt"
+	"strconv"
+	"strings"
+
 	"go.minekube.com/brigodier"
 	. "go.minekube.com/common/minecraft/color"
 	. "go.minekube.com/common/minecraft/component"
 	"go.minekube.com/gate/pkg/command"
-	"go.minekube.com/gate/pkg/internal/suggest"
-	"strconv"
-	"strings"
+	"go.minekube.com/gate/pkg/command/suggest"
 )
 
 const glistCmdPermission = "gate.command.glist"
@@ -24,7 +25,7 @@ func newGlistCmd(proxy *Proxy) brigodier.LiteralNodeBuilder {
 		Then(brigodier.Argument(glistServerArg, brigodier.String).
 			Suggests(command.SuggestFunc(func(_ *command.Context,
 				b *brigodier.SuggestionsBuilder) *brigodier.Suggestions {
-				return suggest.Build(b, b.Input, append(serverNames(proxy), "all"))
+				return suggest.Similar(b, append(serverNames(proxy), "all")).Build()
 			})).
 			Executes(command.Command(func(c *command.Context) error {
 				return glistSendServerCount(proxy, c.Source, c.String(glistServerArg))
